+++ src/test/kotlin/tornadofx/PropertiesTest.kt
@@ -2,6 +2,7 @@ package tornadofx
 
 import org.junit.Test
 import kotlin.test.assertEquals
+import kotlin.test.assertTrue
 
 class PropertiesTest {
 
@@ -42,5 +43,41 @@ class PropertiesTest {
         assertEquals(42, fixture.integer)
         assertEquals(42, fixture.integerDefault)
     }
+    class TestClass {
+        var myProperty: String by singleAssign()
+    }
 
+    @Test
+    fun failNoAssignment() {
+        val instance = TestClass()
+        var failed = false
+
+        try {
+            val extract = instance.myProperty
+        } catch (e: Exception) {
+            failed = true
+        }
+        assertTrue(failed)
+    }
+    @Test
+    fun succeedAssignment() {
+        val instance = TestClass()
+        var failed = false
+        instance.myProperty = "foo"
+        val extract = instance.myProperty
+    }
+    @Test
+    fun failDoubleAssignment() {
+
+        val instance = TestClass()
+        var failed = false
+        instance.myProperty = "foo"
+
+        try {
+            instance.myProperty = "bar"
+        } catch (e: Exception) {
+            failed = true
+        }
+        assertTrue(failed)
+    }
 }