+++ CHANGELOG.md
@@ -4,6 +4,10 @@ All notable changes to this project will be documented in this file.
 ## [Unreleased]
 
 ### Added
+- `singleAssign()` property delegates (https://github.com/edvin/tornadofx/issues/17)
+- Tests for `singleAssign()`
+
+### Added
 - BorderPane builder (https://github.com/edvin/tornadofx/pull/16)
 - `Node.gridpaneConstraints` extension function (https://github.com/edvin/tornadofx/issues/12)
 - `Node.vboxConstraints` extension function